refactor hooks
remove hook service
add processor tests
add datacord hook tests
