+++ apps/api/src/config/configuration.ts
@@ -22,4 +22,10 @@ export default () => ({
       : [],
     methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
   },
+  DATACORD: {
+    PASSWORD: process.env.DATACORD_PASSWORD,
+    USER: process.env.DATACORD_USER,
+    URL: process.env.DATACORD_URL,
+    PAGE_URL: process.env.DATACORD_PAGE_URL,
+  },
 });
