+++ apps/api/src/hook/hook.module.ts
@@ -1,19 +1,13 @@
 import { ConfigModule } from '@nestjs/config';
 import { EmailModule } from '../email/email.module';
-import { EmailProcessor } from './email.processor';
+import { EmailProcessor } from './processors/email.processor';
 import { HookProcessorFactory } from './hook.factory';
-import { HookService } from './hook.service';
 import { Module } from '@nestjs/common';
-import { WebhookProcessor } from './webhook.processor';
+import { WebhookProcessor } from './processors/webhook.processor';
 
 @Module({
   imports: [EmailModule, ConfigModule],
-  providers: [
-    HookService,
-    EmailProcessor,
-    WebhookProcessor,
-    HookProcessorFactory,
-  ],
-  exports: [HookService, HookProcessorFactory],
+  providers: [EmailProcessor, WebhookProcessor, HookProcessorFactory],
+  exports: [HookProcessorFactory],
 })
 export class HookModule {}
